Updates to main demo window
Update to save the Rich Text Box (as xaml package). When an image is loaded via copied / dragged Html, it produces an Image element that then loads the image from the internet. This produces a lag. However, once saved (as xaml package), the image is serialized into the package. And next load goes fast,